+++ hoomd/md/NeighborList.cc
@@ -544,13 +544,6 @@ void NeighborList::checkBoxSize()
     if (m_diameter_shift)
         rmax += m_d_max - Scalar(1.0);
 
-    if (m_filter_body)
-        {
-        // add the maximum diameter of all composite particles
-        Scalar max_d_comp = m_pdata->getMaxCompositeParticleDiameter();
-        rmax += 0.5 * max_d_comp;
-        }
-
     if ((periodic.x && nearest_plane_distance.x <= rmax * 2.0)
         || (periodic.y && nearest_plane_distance.y <= rmax * 2.0)
         || (m_sysdef->getNDimensions() == 3 && periodic.z