+++ hoomd/md/NeighborListGPUStencil.cc
@@ -234,13 +234,6 @@ void NeighborListGPUStencil::buildNlist(uint64_t timestep)
     if (m_diameter_shift)
         rmax += m_d_max - Scalar(1.0);
 
-    if (m_filter_body)
-        {
-        // add the maximum diameter of all composite particles
-        Scalar max_d_comp = m_pdata->getMaxCompositeParticleDiameter();
-        rmax += 0.5 * max_d_comp;
-        }
-
     ArrayHandle<Scalar> d_r_cut(m_r_cut, access_location::device, access_mode::read);
     ArrayHandle<Scalar> d_r_listsq(m_r_listsq, access_location::device, access_mode::read);